refactor: split schema sanitization into own module, promote helpers
- Move sanitize_schema and all schema-related helpers from
  converters/base/tools.py to converters/base/schema.py — schema
  sanitization is a JSON Schema utility, not a tool operations concern
- Re-export sanitize_schema from tools.py for backward compatibility
- Promote check_admin_auth (auth.py) and process_stream_chunk (proxy.py)
  from private helpers to public functions — both are semantically
  distinct concerns suitable for independent testing
