+++ src/llm_rosetta/converters/base/__init__.py
@@ -24,7 +24,8 @@ from .context import ConversionContext, MetadataMode, StreamContext
 from .converter import BaseConverter
 from .messages import BaseMessageOps  # noqa: F401
 from .tool_content import convert_content_blocks_to_ir, convert_ir_content_blocks_to_p  # noqa: F401
-from .tools import BaseToolOps, sanitize_schema  # noqa: F401
+from .schema import sanitize_schema  # noqa: F401
+from .tools import BaseToolOps  # noqa: F401
 
 __all__ = [
     # 主转换器 Main converter
